requirements.txt
numpy==1.26.4
scipy==1.14.0
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
tensorboard==2.17.0
nvidia-smi
nvitop
scikit-learn
labelme2coco
Wand
pytesseract
python-levenshtein==0.25.1
bitsandbytes==0.43.3
accelerate==0.33.0
transformers==4.44.2